[#438] split implementation of the SessionManager
