+++ client/src/commonMain/kotlin/com/github/mustafaozhan/ccc/client/helper/SessionManager.kt
@@ -1,0 +1,7 @@
+package com.github.mustafaozhan.ccc.client.helper
+
+interface SessionManager {
+    fun shouldShowBannerAd(): Boolean
+
+    fun shouldShowInterstitialAd(): Boolean
+}
