+++ client/src/commonMain/kotlin/com/github/mustafaozhan/ccc/client/helper/SessionManagerImpl.kt
@@ -1,16 +1,17 @@
-package com.github.mustafaozhan.ccc.client.util
+package com.github.mustafaozhan.ccc.client.helper
 
+import com.github.mustafaozhan.ccc.client.util.isRewardExpired
 import com.github.mustafaozhan.ccc.common.settings.SettingsRepository
 import com.github.mustafaozhan.config.ConfigManager
 
-class SessionManager(
+class SessionManagerImpl(
     private val configManager: ConfigManager,
     private val settingsRepository: SettingsRepository
-) {
-    fun shouldShowBannerAd() = !settingsRepository.firstRun &&
+) : SessionManager {
+    override fun shouldShowBannerAd() = !settingsRepository.firstRun &&
         settingsRepository.adFreeEndDate.isRewardExpired() &&
         settingsRepository.sessionCount > configManager.appConfig.adConfig.bannerAdSessionCount
 
-    fun shouldShowInterstitialAd() =
+    override fun shouldShowInterstitialAd() =
         settingsRepository.sessionCount > configManager.appConfig.adConfig.interstitialAdSessionCount
 }
