+++ client/src/commonTest/kotlin/com/github/mustafaozhan/ccc/client/viewmodel/CalculatorViewModelTest.kt
@@ -3,8 +3,8 @@
  */
 package com.github.mustafaozhan.ccc.client.viewmodel
 
+import com.github.mustafaozhan.ccc.client.helper.SessionManager
 import com.github.mustafaozhan.ccc.client.mapper.toUIModel
-import com.github.mustafaozhan.ccc.client.util.SessionManager
 import com.github.mustafaozhan.ccc.client.util.after
 import com.github.mustafaozhan.ccc.client.util.before
 import com.github.mustafaozhan.ccc.client.util.getCurrencyConversionByRate
